refactor: update module import path from pkgs/mod to mod
Updated import statements across formula package to use the simplified
module path `github.com/goplus/llar/mod/module` instead of
`github.com/goplus/llar/pkgs/mod/module`. This reflects a restructuring
of the package hierarchy by removing the intermediate `pkgs` directory.

Changes affect:
- formula/classfile.go
- formula/project.go
- formula/project_test.go
